JsonSettingsFile::getOption gets a mandatory default value written if not found in json
This is the way, good old ENEL applications handled settings: When asked for
an option which is not available, store it.

This has the following advantages:

* there is only one common place in GUI (GlobalConfig.xml) where options are
  added/removed. Default configuration shipped can be reduced to minimum.
* in case new options are added we HAVE to think about their defaults

To avoid QML warnings "Binding loop detected for property..." the signal
settingsChanged was split into settingsSaveRequest and settingsChanged and
settingsChanged is fired only in case the setting really change.

As nice side effect of this change (together with modifications in GUI), the
json file is written only for:

* options read for the first time
* options modified

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>